+++ src/connection/Connection.ts
@@ -1,6 +1,11 @@
 import JsonRPCError from "../error/JsonRPCError.js";
 import {EventEmitter} from "eventemitter3";
 import ConnectionEventData from "./ConnectionEventData";
+import InvalidResponseError from "../error/InvalidResponseError";
+
+export type SuccessfulCallResponse = {success: true, data: unknown};
+export type FailedCallResponse = {success: false, error: unknown};
+export type CallResponse = SuccessfulCallResponse | FailedCallResponse;
 
 /**
  * Library independent connection interface.
@@ -13,7 +18,7 @@ export default abstract class Connection extends EventEmitter<ConnectionEventDat
      * @param parameters The parameters to pass to the method.
      * @returns A promise that resolves to the result of the method call.
      */
-    protected abstract callRaw(method: string, parameters: object | Array<unknown>): Promise<unknown>;
+    protected abstract callRaw(method: string, parameters: object | Array<unknown>): Promise<CallResponse>;
 
     /**
      * Call a method on the server with the given parameters. If an error occurs, it throws an error.
@@ -23,11 +28,23 @@ export default abstract class Connection extends EventEmitter<ConnectionEventDat
      * @throws JsonRPCError if the server returns an error.
      */
     public async call(method: string, parameters: object | Array<unknown>): Promise<unknown> {
-        const data = await this.callRaw(method, parameters);
-        if (typeof data === 'object' && data && 'error' in data && typeof data.error === 'object' && data.error) {
-            throw JsonRPCError.parse(data.error, "error");
+        const response = await this.callRaw(method, parameters);
+        if (response.success) {
+            return response.data;
         }
-        return data;
+
+        let error = response.error;
+        if (typeof error === "object" && error) {
+            try {
+                error = JsonRPCError.parse(error, "error");
+            } catch (e) {
+                if (!(e instanceof InvalidResponseError)) {
+                    throw e;
+                }
+            }
+        }
+
+        throw error;
     }
 
     /**