+++ test/TestConnection.ts
@@ -1,4 +1,5 @@
 import {Connection} from "../src";
+import {CallResponse} from "../src/connection/Connection";
 
 export type RequestHistoryEntry = {
     method: string,
@@ -6,11 +7,16 @@ export type RequestHistoryEntry = {
 }
 
 export default class TestConnection extends Connection {
-    private responseQueue: unknown[] = [];
+    private responseQueue: CallResponse[] = [];
     private requestHistory: RequestHistoryEntry[] = [];
 
-    public addResponse<T>(result: T): T {
-        this.responseQueue.push(result);
+    public addSuccess<T>(result: T): T {
+        this.responseQueue.push({success: true, data: result});
+        return result;
+    }
+
+    public addError<T>(result: T): T {
+        this.responseQueue.push({success: false, error: result});
         return result;
     }
 
@@ -18,13 +24,14 @@ export default class TestConnection extends Connection {
         return this.requestHistory.shift() ?? null;
     }
 
-    async callRaw(method: string, parameters: object | Array<unknown>): Promise<unknown> {
+    async callRaw(method: string, parameters: object | Array<unknown>): Promise<CallResponse> {
         this.requestHistory.push({method, parameters});
-        if (this.responseQueue.length === 0) {
+        const result = this.responseQueue.shift();
+        if (result === undefined) {
             throw new Error("No more results in queue");
         }
 
-        return this.responseQueue.shift();
+        return result;
     }
 
     close(): void {