+++ test/error.test.ts
@@ -1,5 +1,6 @@
 import {expect, test} from "vitest";
-import {JsonRPCError, MissingPropertyError, IncorrectTypeError} from "../src";
+import {JsonRPCError, MissingPropertyError, IncorrectTypeError, JsonRPCErrorCode} from "../src";
+import {getConnection} from "./utils";
 
 test("Parse error missing code", () => {
     const data = {message: "Error message"};
@@ -27,4 +28,16 @@ test("Parse error wrong type message", () => {
     expect(() => JsonRPCError.parse(data)).toThrow(
         new IncorrectTypeError("string", "number", data, "message")
     );
+});
+
+test('Test method not found', async () => {
+    const connection = await getConnection();
+    await expect(() => connection.call("aternos:invalid-method", {}))
+        .rejects.toThrow(new JsonRPCError(JsonRPCErrorCode.METHOD_NOT_FOUND, "Method not found", "Method not found: aternos:invalid-method"));
+});
+
+test('Test invalid params', async () => {
+    const connection = await getConnection();
+    await expect(() => connection.call("minecraft:operators/add", {}))
+        .rejects.toThrow(new JsonRPCError(JsonRPCErrorCode.INVALID_PARAMS, "Invalid params", "Params passed by-name, but expected param [add] does not exist"));
 });